+++ contracts/StreamEmployee.sol
@@ -5,14 +5,14 @@ pragma solidity ^0.6.0;
 import "@openzeppelin/contracts/math/SafeMath.sol";
 
 contract StreamEmployee {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     address public employeeAddress;
-    uint public payPerHour;
+    uint256 public payPerHour;
     bool public isWorking;
-    uint public workStarted;
+    uint256 public workStarted;
 
-    uint public balance;
+    uint256 public balance;
 
     modifier workHasNotBegun() {
         require(false == isWorking, "Work has already been started");
@@ -24,7 +24,7 @@ contract StreamEmployee {
         _;
     }
 
-    constructor(uint _payPerHour, address _employeeAddress) public {
+    constructor(uint256 _payPerHour, address _employeeAddress) public {
         payPerHour = _payPerHour;
         employeeAddress = _employeeAddress;
     }
@@ -48,7 +48,7 @@ contract StreamEmployee {
         balance = balance.add(payAccrued());
     }
 
-    function payAccrued() public view returns (uint) {
+    function payAccrued() public view returns (uint256) {
         if (timeWorkedInSeconds() > 0) {
             return timeWorkedInSeconds().mul(payPerSecond());
         }
@@ -56,13 +56,13 @@ contract StreamEmployee {
         return 0;
     }
 
-    function payPerSecond() public view returns (uint) {
-        uint hourly = payPerHour.mul(10 ** 18);
-        uint minutely = hourly.div(60);
+    function payPerSecond() public view returns (uint256) {
+        uint256 hourly = payPerHour.mul(10**18);
+        uint256 minutely = hourly.div(60);
         return minutely.div(60);
     }
 
-    function timeWorkedInSeconds() public view returns (uint) {
+    function timeWorkedInSeconds() public view returns (uint256) {
         return block.timestamp.sub(workStarted);
     }
 }
